+++ utils/core/csvToHtml_devices.sh
@@ -20,7 +20,7 @@
 #
 #
 # Author..............: Pierre-Yves Lapersonne
-# Version.............: 16.0.0
+# Version.............: 17.0.0
 # Since...............: 18/08/2016
 # Description.........: Process a file/an input (mainly in CSV format) to HTML with CSS if needed
 #			This file must contain several columns: Type, OS, Constructor, Name, Screen size, Sreen type, Screen resolution, SoC, GPU, Sensors, Batery, Storage, RAM, Camera, Dimensions, Weight, IP, USB Type, SD Card, SIM , UI
@@ -68,10 +68,12 @@ while read -r line; do
 		if [ $currentRowIndex -eq $(($NUMBER_OF_LINES_TO_IGNORE - 1)) ]; then
 			echo -e "\t<thead>"
 			echo -e "\t\t<tr>"
-			# For GNU/Linux (good and best) systems
-			#echo $line | sed 's/;/\n/g' | while read -r item; do
-			# For macOS (not so best) systems	
-			echo $line | sed 's/;/\'$'\n/g' | while read -r item; do
+			if [ $(DoesRunOnGNULinux) == "yes" ]; then # GNU/Linux
+				regex="s/;/\n/g"
+			else # macOS
+				regex="s/;/\'$'\n/g"
+			fi
+			echo $line | sed $regex | while read -r item; do
 				echo -e "\t\t\t<td class=\"header\">" $item "</td>"
 			done
 			echo -e "\t\t</tr>"
@@ -85,11 +87,13 @@ while read -r line; do
 	echo -e "\t\t<tr>"
 
 	# ***** Step 4: Split the line and replace ; by \n, and delete useless "
-	fieldIndex=0;
-	# For GNU/Linux (good and best) systems
-	#echo $line | sed 's/;/\n/g' | while read -r item; do
-	# For macOS (not so best) systems	
-	echo $line | sed 's/;/\'$'\n/g' | while read -r item; do
+	fieldIndex=0
+	if [ $(DoesRunOnGNULinux) == "yes" ]; then # GNU/Linux
+		regex="s/;/\n/g"
+	else # macOS
+		regex="s/;/\'$'\n/g"
+	fi
+	echo $line | sed $regex | while read -r item; do
 		cleanItem=`echo $item | sed 's/\"//g'`
 		# Add an good CSS class
 		case "$fieldIndex" in