+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 ### Changes
 
 - [#15](https://github.com/pylapp/Tips-tools/issues/15) Add dry-run script 
+- [#10](https://github.com/pylapp/Tips-tools/issues/10) Deal with GNU/Linux and macOS primitives
 
 ## v14.3.3